mfcc_app/requirements.txt
Flask
Flask-Cors
aiokafka
boto3
botocore
pydantic
python-dotenv
flasgger